+++ decidim-core/spec/commands/decidim/create_omniauth_registration_spec.rb
@@ -86,6 +86,28 @@ module Decidim
             expect(user.valid_password?("abcde1234")).to eq(true)
           end
 
+          it "notifies about registration with oauth data" do
+            user = create(:user, email: email, organization: organization)
+            identity = Decidim::Identity.new(id: 1234)
+            expect(command).to receive(:create_identity).and_return(identity)
+
+            expect(ActiveSupport::Notifications)
+              .to receive(:publish)
+              .with(
+                "decidim.events.user.omniauth_registration",
+                user_id: user.id,
+                identity_id: 1234,
+                provider: provider,
+                uid: uid,
+                email: email,
+                name: "Facebook User",
+                nickname: "facebook_user",
+                avatar_url: "http://www.example.com/foo.jpg",
+                raw_data: {}
+              )
+            command.call
+          end
+
           describe "user linking" do
             context "with a verified email" do
               let(:verified_email) { email }